Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/svg/painting/parsing/text-rendering-valid-expected.txt


PASS e.style['text-rendering'] = "auto" should set the property value
FAIL e.style['text-rendering'] = "optimizespeed" should set the property value assert_equals: serialization should be canonical expected "optimizespeed" but got "optimizeSpeed"
FAIL e.style['text-rendering'] = "optimizelegibility" should set the property value assert_equals: serialization should be canonical expected "optimizelegibility" but got "optimizeLegibility"
FAIL e.style['text-rendering'] = "geometricprecision" should set the property value assert_equals: serialization should be canonical expected "geometricprecision" but got "geometricPrecision"
